Updating App.jsx file
