+++ src/App.jsx
@@ -1,23 +1,34 @@
-import "./App.css";
+import React from "react";
 
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://pokeapi.co/"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Gotta Catch'Em All!!!!
-        </a>
-      </header>
-    </div>
-  );
-}
+import {BrowserRouter, Route, Routes} from "react-router-dom";
+import Pokemon from "./pages/Pokemon.js";
+import Home from "./pages/Home";
+
+import NotFound from "./pages/NotFound";
+import Layout from "./pages/Layout.js";
+
+import {createTheme, ThemeProvider} from "@mui/material/styles";
+import {getDesignTokens} from "./resources/theme.js";
+
+
+const App = () => {
+
+
+    const theme = React.useMemo(() => createTheme(getDesignTokens('light')), []);
+
+    return (
+        <ThemeProvider theme={theme}>
+            <BrowserRouter>
+                <Layout>
+                    <Routes>
+                        <Route path="/" element={<Home/>}/>
+                        <Route path="pokemon/:pokemonId" element={<Pokemon/>}/>
+                        <Route path="*" element={<NotFound/>}/>
+                    </Routes>
+                </Layout>
+            </BrowserRouter>
+        </ThemeProvider>
+    );
+};
 
 export default App;
